Split out a renderer module
