+++ js/controlbox.js
@@ -1,3 +1,5 @@
+var d3 = require('d3');
+
 "use strict";
 
 /**
